backend/pyproject.toml
[project]
name = "class-scheduler-backend"
version = "0.1.0"
description = "POC backend for class scheduling with Zoom integration"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "python-dateutil>=2.9",
  "tzdata>=2024.1",
  "sqlalchemy>=2.0.48",
  "aiosqlite>=0.22.1",
  "greenlet>=3.3.2",
  "google-api-python-client>=2.192.0",
  "google-auth-httplib2>=0.3.0",
  "google-auth-oauthlib>=1.3.0",
  "aiogoogle>=5.17.0",
]

[tool.uv]
package = false

[project.scripts]
dev = "uvicorn:main"
